backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
python-multipart>=0.0.6
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
pymupdf>=1.23.3
python-docx>=0.8.11
beautifulsoup4>=4.12.2
youtube-transcript-api>=0.6.1
requests>=2.31.0
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.0
lxml>=4.9.3
aiofiles>=23.2.1
pydantic>=2.0.0
pydantic-settings>=2.0.0
trafilatura>=1.6.1
readability-lxml>=0.8.1
html2text>=2020.1.16
unicodedata2>=15.0.0
